requirements.txt
flask
flask-restful
flask-jwt-extended
flask-sqlalchemy
flask-migrate
flask-cors

marshmallow
marshmallow-sqlalchemy

redis
gunicorn
werkzeug

python-dotenv